+++ src/Game/Werewolf/Engine.hs
@@ -127,7 +127,7 @@ checkStage' = use stage >>= \stage' -> case stage' of
                     ifM (uses protect $ maybe False (== target ^. name))
                         (events %= cons (ProtectEvent $ target ^. name))
                         (events %= cons (DevourEvent $ target ^. name))
-                _           -> tell [noPlayerDevouredMessage]
+                _           -> events %= cons NoDevourEvent
 
             protect .= Nothing
 
@@ -169,8 +169,9 @@ checkEvents = do
     mapM_ applyEvent available
 
 eventAvailable :: MonadState Game m => Event -> m Bool
-eventAvailable (DevourEvent _) = gets isSunrise
-eventAvailable (PoisonEvent _) = gets isSunrise
+eventAvailable (DevourEvent _)  = gets isSunrise
+eventAvailable NoDevourEvent    = gets isSunrise
+eventAvailable (PoisonEvent _)  = gets isSunrise
 eventAvailable (ProtectEvent _) = gets isSunrise
 
 applyEvent :: (MonadState Game m, MonadWriter [Message] m) => Event -> m ()
@@ -179,19 +180,20 @@ applyEvent (DevourEvent targetName) = do
     heal'   <- use heal
 
     if heal'
-        then tell [playerHealedMessage $ player ^. name]
+        then tell [noPlayerDevouredMessage]
         else do
             killPlayer player
             tell [playerDevouredMessage player]
 
     heal .= False
-applyEvent (PoisonEvent name) = do
+applyEvent NoDevourEvent            = tell [noPlayerDevouredMessage]
+applyEvent (PoisonEvent name)       = do
     player <- uses players $ findByName_ name
 
     killPlayer player
 
     tell [playerPoisonedMessage player]
-applyEvent (ProtectEvent name) = tell [playerProtectedMessage name]
+applyEvent (ProtectEvent name)      = tell [playerProtectedMessage name]
 
 checkGameOver :: (MonadState Game m, MonadWriter [Message] m) => m ()
 checkGameOver = do